backend/requirements.txt
bootstrap-icons
python-dotenv
requests